INFO-Capstone-Project-2026-Cursor-IPO-Newsletter-ipo-hype-tracker/python-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytrends==4.9.2
yfinance==0.2.18
requests==2.31.0
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.24.3
python-multipart==0.0.6
asyncpg==0.29.0
